Moved Man in the middle to its own folder
