+++ InfernalRobotics/InfernalRobotics/Control/ILinearControl.cs
@@ -1,0 +1,6 @@
+namespace InfernalRobotics.Control
+{
+    public interface ILinearControl
+    {
+    }
+}
