+++ InfernalRobotics/InfernalRobotics/Extension/ServoExtension.cs
@@ -1,4 +1,5 @@
 ﻿using InfernalRobotics.Command;
+using InfernalRobotics.Control;
 using InfernalRobotics.Module;
 using System.Collections.Generic;
 using System.Linq;
